backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-speech==2.21.0
google-cloud-storage==2.13.0
numpy==1.23.5
# For Apple Silicon (M1/M2/M3)
tensorflow-macos==2.12.0
tensorflow-metal==0.8.0
tensorflow-hub==0.14.0
opencv-python==4.8.0.74
pydub==0.25.1
pillow==10.0.0
scikit-learn==1.3.2
scipy==1.11.3
matplotlib==3.8.0 